transcription-api/requirements.txt
fastapi
faster-whisper>=0.10.0
python-dotenv
uvicorn
python-multipart
ffmpeg-python
soundfile
groq
supabase
boto3
pyannote.audio>=3.1.1
torch
torchaudio
